[RAPPS] move settings-related stuff to a separate file
